requirements.txt
python-binance>=1.0.28
python-dotenv==1.0.0
tenacity==8.2.3